+++ pkg/key/key.go
@@ -4,5 +4,5 @@ const (
 	DNSFinalizerNameOld = "dns-operator-openstack.finalizers.giantswarm.io"
 	DNSFinalizerNameNew = "dns-operator-route53.finalizers.giantswarm.io"
 
-	AnnotationWildcardCNAMETarget = "dns-operator-route53.giantswarm.io/wildcard-cname-target"
+	AnnotationWildcardCNAMETarget = "network.giantswarm.io/wildcard-cname-target"
 )